pyproject.toml
[tool.poetry]
name = "498final"
version = "0.1.0"
description = ""
authors = ["Adithya Yerramsetty <adithya.yerramsetty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^9.5.0"
imageio = "^2.28.1"
av = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
